% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.barchart.R
\name{acs.barchart}
\alias{acs.barchart}
\title{Make barchart from ACS PUMS data.}
\usage{
acs.barchart(data, x, stack = NULL, title = NULL, xlab = NULL,
  ylab = NULL)
}
\arguments{
\item{data}{ACS PUMS data for housing or population}

\item{x}{(string) name of categorical variable in data to put on x-axis}

\item{stack}{optional (string) name of categorical value for stacked bar chart}

\item{title}{optional plot title}

\item{xlab}{optional x label}

\item{ylab}{optional y label}
}
\value{
ggplot of weighted count data
}
\description{
Make a barchart of weighted counts within categories from the ACS PUMS data.
Optionally, this can be a stacked barchart, partitioned over two variables.
}
\examples{
wa.house16$Tenure <- own.rent(wa.house16)
acs.barchart(wa.house16, 'Tenure',
             title='Housing Unit Count by Tenure', xlab='Tenure')

wa.pop16$sex <- acs.recode(wa.pop16, 'SEX', data.dict16)
wa.pop16$emp.stat <- acs.translate(wa.pop16, 'ESR', c(NA, 1:6),
    c('NILF', 'Employed', 'Employed', 'Unemployed', 'Employed', 'Employed', 'NILF'))
acs.barchart(wa.pop16, 'emp.stat', stack='sex',
             title='Employment Status by Gender',
             xlab='Employment Status')

}
