% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.boxplot.R
\name{acs.boxplot}
\alias{acs.boxplot}
\title{Make boxplot from ACS PUMS data}
\usage{
acs.boxplot(data, x, y, title = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{data}{ACS PUMS data for housing or population}

\item{x}{(string) name of categorical variable in data to put on x-axis}

\item{y}{(string) name of numeric variable in data for y-axis}

\item{title}{optional plot title}

\item{xlab}{optional x label}

\item{ylab}{optional y label}
}
\value{
ggplot of the weighted distribution of y, partitioned by x
}
\description{
Make a weighted boxplot of the distribution of a numeric field in the
ACS PUMS data, partitioned by a categorical field.
}
\examples{
wa.house16$Beds <- clip.column(wa.house16, 'BDSP', 5)
acs.boxplot(wa.house16[!is.na(wa.house16$GRNTP),], 'Beds', 'GRNTP',
            title='Rent vs. Bedrooms', xlab='Beds (max. 5)',
            ylab='Gross Rent($)')

}
