% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.cut.R
\name{acs.cut}
\alias{acs.cut}
\title{Cut numeric column and assign labels to bins}
\usage{
acs.cut(x, field, cuts, labels = NULL)
}
\arguments{
\item{x}{input data frame}

\item{field}{name of numeric column in x to cut}

\item{cuts}{cut points. If \code{cuts} does not cover the range of the data,
then extra bins will be returned on the ends.}

\item{labels}{optional character vector of level labels.
If used, must be same length as the factor produced.}
}
\value{
factor with binned values for \code{x[[field]]}
        and the names from \code{labels}
}
\description{
Cuts a numeric column using \code{Hmisc::cut2} and optionally assigns different
labels to the levels produced.
If labels are provided, there must be as many of them as there are levels.
}
\examples{
med.inc <- acs.median(wa.house16, 'HINCP')
cuts <- med.inc * c(0.3, 0.5, 0.8, 1.2)
labels <- c('0-30 AMI', '30-50 AMI', '50-80 AMI', '80-120 AMI', '>120 AMI')
wa.house.16$Income.Band <- acs.cut(wa.house16, 'HINCP', cuts=cuts, labels=labels)
group.count(wa.house16, 'Income.Band', drop.na.group=TRUE)

}
