% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.describe.R
\name{acs.describe}
\alias{acs.describe}
\title{Describe a field in the ACS PUMS data}
\usage{
acs.describe(dd, field, max.len = NULL)
}
\arguments{
\item{dd}{the data dict as a string}

\item{field}{the ACS column name to extract}

\item{max.len}{max length of description in words, default is all}
}
\value{
list with column name and description, levels and their definitions
}
\description{
Extract the description of a field in the ACS PUMS data from the
data dictionary. The description includes the column name (abbreviation),
the field name (plain English), the factor levels and their descriptions
for integer-coded categoricals or the range and description for numeric
variables.
}
\examples{
# Integer-coded categorical field
acs.describe(data.dict16, 'TEN')

# Numeric field
acs.describe(data.dict16, 'NP', 25)

}
