% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.group.estimate.R
\name{acs.group.estimate}
\alias{acs.group.estimate}
\title{Compute count statistics on grouped ACS PUMS data.}
\usage{
acs.group.estimate(x, gp_var, include.total = TRUE)
}
\arguments{
\item{x}{a data frame of PUMS data}

\item{gp_var}{name of variable in x to group data by, a string}

\item{include.total}{include the total across all groups as last row,
default TRUE}
}
\value{
data frame of count, standard error of the count,
        and the margin of error of the count, over groups
        defined by \code{gp_var} and the whole data.
}
\description{
Collect point estimate, standard error and 90% margin-of-error for
the count across groups defined by \code{gp_var} and (optionally)
over the whole data.
}
\examples{
acs.group.estimate(wa.house16, 'TEN')

}
