% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.levels.R
\name{acs.levels}
\alias{acs.levels}
\title{Map integer levels to their descriptions}
\usage{
acs.levels(dd, token, max.len = -1)
}
\arguments{
\item{dd}{the data dict as a string}

\item{token}{the ACS short name to extract}

\item{max.len}{max length of description in characters, default is all}
}
\value{
data frame from the integer level code to the level description
}
\description{
Builds a data frame mapping integer coded categorical variable levels
in the ACS PUMS data to a factor named with text from the data dictionary.
}
\examples{
library(readr)
# Get file path for data dict, which is in inst/extdata/ in the package
dd.file <- system.file('extdata', 'PUMS_Data_Dictionary_2016.txt',
               package='PUMSutils', mustWork=TRUE)
dd <- read_file(dd.file)
acs.levels(dd, 'ESR', max.len = 30)

}
