% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.line.estimate.R
\name{acs.line.estimate}
\alias{acs.line.estimate}
\title{Compute count statistics from ACS PUMS data.}
\usage{
acs.line.estimate(x)
}
\arguments{
\item{x}{a data frame of PUMS data, filtered to the in-sample rows.}
}
\value{
the point estimate, standard error and 90% margin of error
for the size of this sample.
}
\description{
Compute point estimate, standard error and 90% margin of error
on ACS PUMS data using the replicate weights and return results on one line.
}
\examples{
acs.line.estimate(wa.house16)

}
