% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.mean.R
\name{acs.mean}
\alias{acs.mean}
\title{Estimates means from the ACS PUMS data.}
\usage{
acs.mean(x, field, wt.rep.num = NULL)
}
\arguments{
\item{x}{data.frame or similar with data filtered to relevant rows}

\item{field}{the data field to take the mean over}

\item{wt.rep.num}{the weight replicate number, if a replicate weight}
}
\value{
an estimate of the mean of the data field over this data
}
\description{
Estimates the mean of the data field over the population 
represented by a sample drawn from ACS PUMS data.
}
\note{
some ACS fields are top-coded, so their means are not correct, 
check the data dictionary or the PUMS Readme for more.
}
\references{
\url{http://www2.census.gov/programs-surveys/acs/tech_docs/pums/ACS2013_PUMS_README.pdf}
}
