% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.quantile.R
\name{acs.quantile}
\alias{acs.quantile}
\title{Quantiles for ACS PUMS Data.}
\usage{
acs.quantile(x, field, probs = c(0.1, 0.25, 0.5, 0.75, 0.9),
  wt.rep.num = NULL)
}
\arguments{
\item{x}{data.frame or similar with data filtered to relevant rows}

\item{field}{the data field to take the quantiles over}

\item{probs}{a vector of quantiles to take}

\item{wt.rep.num}{the weight replicate number, if a replicate weight}
}
\value{
a vector of quantiles for the selected field over the given data
}
\description{
Estimate a range of quantiles on the ACS PUMS Data, using sample weights.
}
