% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.recode.R
\name{acs.recode}
\alias{acs.recode}
\title{Rewrite variable levels using data dictionary}
\usage{
acs.recode(x, field, dd, max.len = -1)
}
\arguments{
\item{x}{ACS PUMS data frame}

\item{field}{column name for a variable in the ACS PUMS data}

\item{dd}{the ACS data dictionary loaded as a string}

\item{max.len}{max length of description in characters, default is all}
}
\value{
factor column with integer levels in \code{x[field]} replaced
        with the desciption for that level from the data dictionary.
}
\description{
In the ACS PUMS data, many fields are integer-coded categoricals.
There is a data dictionary (a text document) that describes the mapping
from integers to level definitions. Function \code{rename.levels} takes
the data dicitonary as a string and uses it to construct a factor variable
that recodes integer levels to the text definitions given in the data
dictionary.
}
