% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.table.R
\name{acs.table}
\alias{acs.table}
\title{Compute a Table on ACS PUMS Data.}
\usage{
acs.table(x, field)
}
\arguments{
\item{x}{data.frame or similar with data filtered to relevant rows}

\item{field}{name of the field to tabulate}
}
\value{
table of weighted counts of the values of field in x
}
\description{
Computes a table for a field in the ACS PUMS data, using the sample weights.
}
