% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.total.R
\name{acs.total}
\alias{acs.total}
\title{Compute a weighted total of a numerical field in ACS PUMS data.}
\usage{
acs.total(x, field, wt.rep.num = NULL)
}
\arguments{
\item{x}{data.frame or similar with data filtered to relevant rows}

\item{field}{the data field to take the median of}

\item{wt.rep.num}{the weight replicate number, if a replicate weight}
}
\value{
an estimate of the weighted total of field in the data
}
\description{
Computes an estimate of a total of a numerical field in the
ACS PUMS data using the correct sample weights. Note that these
values may be slightly in error due to top or bottom coding of some
numeric variables.
}
\examples{
# Estimated total rent paid by reters in Washington State in 2016.
12 * acs.total(wa.house16, 'GRNTP')

}
