% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.translate.R
\name{acs.translate}
\alias{acs.translate}
\title{Rewrite values in ACS PUMS data}
\usage{
acs.translate(x, field, old, new, level.order = NULL)
}
\arguments{
\item{x}{a data frame of PUMS data}

\item{field}{string name of the field in x to translate}

\item{old}{vector of the unique values in \code{x$field}}

\item{new}{vector of replacement values (strings).
Must be same length as \code{old}, but values do not have
to be unique (see example).}

\item{level.order}{order to give to factor levels, if used.
Default NULL uses ordering of \code{unique(new)}}
}
\value{
factor vector of same length as \code{x$field} with values from
        \code{new} substituted for the originals
}
\description{
Many variables in the ACS PUMS data are integer-coded categorical values.
For readability, we often want to replace these with strings or factors.
\code{acs.translate} takes the values in the variable \code{field} and
substitutes elements of \code{new} for the corresponing element of \code{old}.
The result is a factor with the levels in the order of \code{level.order} if used,
or in the order of \code{unique(new)} otherwise.
}
\examples{
new.vals <- c('own','own','rent', 'rent', 'vacant')
wa.house16$Tenure <- acs.translate(wa.house16, 'TEN', c(1:4, NA), new.vals)
acs.group.estimate(wa.house16, 'Tenure')

# Strings would not be in this order
in.order <- c('vacant', 'own', 'rent')
wa.house16$Tenure <- acs.translate(wa.house16, 'TEN', c(1:4, NA),
                                   new.vals, level.order=in.order)
acs.group.estimate(wa.house16, 'Tenure')

}
