% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.column.R
\name{clip.column}
\alias{clip.column}
\title{Clip an integer column}
\usage{
clip.column(x, colname, clip.at, clip.name = NULL)
}
\arguments{
\item{x}{a data frame of PUMS data}

\item{colname}{(string) name of column in x to clip}

\item{clip.at}{upper clip level}

\item{clip.name}{(string) name of the top clip level;
for \code{clip.at} of n, the default is "n+".}
}
\value{
factor column with levels like "1", "2", ... "n+" in place
        of the (count) integer values in \code{x$colname}
}
\description{
Many count variables in the ACS PUMS data have long upper tails.
This function clips integer columns and returns a factor with
level names like "1", "2", "3+" that order correctly, even if the clip level
is named something like ">=3", which would not sort correctly as a string.
}
\examples{
wa.house16$Size <- clip.column(wa.house16, 'NP', 5, '>=5')
acs.group.estimate(wa.house16, 'Size')

}
