% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate Count or Total Using ACS PUMS Data.}
\usage{
estimate(x, wt.rep.num = NULL)
}
\arguments{
\item{x}{data.frame or similar with data filtered to relevant rows.}

\item{wt.rep.num}{the weight replicate number, if a replicate weight}
}
\value{
an estimate of the population total for this data.
}
\description{
Estimates a total of the population or sub-group in this data.
}
\references{
\url{http://www2.census.gov/programs-surveys/acs/tech_docs/pums/ACS2013_PUMS_README.pdf}
}

