% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.df.R
\name{extract.field}
\alias{extract.field}
\title{Extract text from data dictionary for a variable}
\usage{
extract.field(dd, token)
}
\arguments{
\item{dd}{the data dict as a string}

\item{token}{the ACS short name to extract}
}
\value{
character vector with part of field, one element per line
}
\description{
Extract all of the text after a token.
The token is the ACS column name. It will be in all caps.
It will start the first line of the field definition.
}
