% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.df.R
\name{fmt.chunk.str}
\alias{fmt.chunk.str}
\title{Formats list of chunks from get.chunk.str}
\usage{
fmt.chunk.str(s)
}
\arguments{
\item{s}{a string with one chunk from get.chunk.str}
}
\value{
character vector with one line from data dictionary per element
}
\description{
Formats list of chunks from get.chunk.str
}
