% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.df.R
\name{get.chunk.str}
\alias{get.chunk.str}
\title{Splits data dict into chunks separated by blank lines}
\usage{
get.chunk.str(dd)
}
\arguments{
\item{dd}{the data dictionary as a string}
}
\value{
list of all chunks in the data dictionary
}
\description{
Splits data dict into chunks separated by blank lines
}
