% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.weight.R
\name{get.weight}
\alias{get.weight}
\title{Get Expansion Weight From ACS PUMS Data.}
\usage{
get.weight(data, wt.rep.num = NULL)
}
\arguments{
\item{data}{ACS housing or population data frame}

\item{wt.rep.num}{The weight replicate number, if a replicate weight
default is NULL for a point estimate}
}
\value{
The correct weight name in this data.
}
\description{
Extracts an expansion weight or weight replicate from either ACS housing 
or population data, regardless of case (which varies).
Fails with error if the requested weight is not present.
}

