% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.estimate.R
\name{group.estimate}
\alias{group.estimate}
\title{Compute estimates and standard errors on grouped ACS PUMS data.}
\usage{
group.estimate(result.name, f, x, gp.var, ..., drop.na.group = FALSE,
  include.total = TRUE)
}
\arguments{
\item{result.name}{name of estimate column in result}

\item{f}{a function to calculate the statistic.
It must take data and a weight replicate number called wt.rep.num
with a default value of NULL.}

\item{x}{a data frame of PUMS data}

\item{gp.var}{name of variable in x to group data by, a string}

\item{...}{other data passed to f}

\item{drop.na.group}{default FALSE, drop the group where `gp.var` is NA}

\item{include.total}{include the total across all groups, default TRUE}
}
\value{
data frame of estimate, standard error of the estimate,
        and the margin of error of the estimate, for the statistic
        defined by `f` , over the groups defined by \code{gp.var}
        and (optionally) the whole data.
}
\description{
Collect point estimate, standard error and 90% margin-of-error for
any statistic across groups defined by \code{gp.var} and (optionally)
over the whole data.
}
\examples{
# Total of occupied households in Washington State in 2016
group.estimate('HH.Ct', estimate, wa.house16, 'TEN', drop.na.group=TRUE)

# 90\% household income by tenure type for Washington State in 2016
group.estimate('HH.Inc90', acs.quantile, wa.house16, 'TEN',
               field='HINCP', probs=0.9, drop.na.group = TRUE)

}
