% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.median.R
\name{group.median}
\alias{group.median}
\title{Compute weighted median of numerical column on grouped ACS PUMS data.}
\usage{
group.median(x, result.name, gp.var, med.var, drop.na.group = TRUE)
}
\arguments{
\item{x}{a data frame of PUMS data}

\item{result.name}{name of estimate column in result}

\item{gp.var}{name of variable in x to group data by, a string}

\item{med.var}{(string) name of a numerical column in x to take median over}

\item{drop.na.group}{drop the group where \code{gp.var} is NA; default TRUE}
}
\value{
data frame of median, standard error of the median,
        and the margin of error of the median of \code{med.var}
        over the groups defined by \code{gp.var} and for the whole data.
}
\description{
Collect point estimate, standard error and 90% margin-of-error for
the median of \code{med.var} across groups defined by \code{gp.var} and
over the whole data.
}
\examples{
# Median income by tenure type for Washington State in 2016
group.median(wa.house16, 'Med.HH.Inc', 'TEN', 'HINCP', drop.na.group=TRUE)

}
