% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.total.R
\name{group.total}
\alias{group.total}
\title{Compute aggregate of numerical column on grouped ACS PUMS data.}
\usage{
group.total(x, result.name, gp.var, tot.var, drop.na.group)
}
\arguments{
\item{x}{a data frame of PUMS data}

\item{result.name}{name of estimate column in result}

\item{gp.var}{name of variable in x to group data by, a string}

\item{tot.var}{(string) name of a numerical column in `x` to aggregate}

\item{drop.na.group}{drop the group where `gp.var` is NA}
}
\value{
data frame of aggregate, standard error of the aggregate,
        and the margin of error of the aggregate of `tot.var`
        over the groups defined by `gp.var` and for the whole data.
}
\description{
Collect point estimate, standard error and 90% margin-of-error for
aggregate of `tot.var` across groups defined by `gp.var` and
over the whole data.
}
\examples{
# Total household population of Washington State in 2016
group.total(x, 'HH.Pop', 'TEN', 'NP', drop.na.group=TRUE)

}
