% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh.type.R
\name{hh.type}
\alias{hh.type}
\title{Classify households by type}
\usage{
hh.type(house)
}
\arguments{
\item{house}{ACS housing data or subset}
}
\value{
character vector of household type
}
\description{
Classifies each household as vacant, one person, a couple, a family with
children present, or a composite household (none of the above).
Note that \code{house} has to be full ACS PUMS housing data, not the
subset of columns used in \code{wa.house16}.
}
