% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.df.R
\name{levels.df}
\alias{levels.df}
\title{Map integer levels to their descriptions}
\usage{
\method{levels}{df}(dd, token)
}
\arguments{
\item{dd}{the data dict as a string}

\item{token}{the ACS short name to extract}
}
\value{
data frame from the integer level code to the level description
}
\description{
Builds a data frame mapping integer coded categorical variable levels
in the ACS PUMS data to a factor named with text from the data dictionary.
}
