% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.estimate.R
\name{line.estimate}
\alias{line.estimate}
\title{Compute an estimate and standard error from ACS PUMS data.}
\usage{
line.estimate(x, f, ..., result.name = NULL)
}
\arguments{
\item{x}{a data frame of PUMS data.}

\item{f}{a function to calculate the statistic.
It must take data and a weight replicate number called
\code{wt.rep.num} with a default value of NULL.}

\item{...}{other data passed to f}

\item{result.name}{name of estimate column in result.
Default of NULL uses \code{deparse(substitute(f))}.}
}
\value{
the point estimate, standard error and 90% margin of error
for the size of this sample.
}
\description{
Compute point estimate, standard error and 90% margin of error
of any statistic on ACS PUMS data using the direct method,
involving replicate weights and return results on one line.
}
\examples{
# Number of households in Washington State in 2016
line.estimate(wa.house16, estimate)

# Fraction of Washington State households that rent for cash
line.estimate(subset(wa.house16, TEN==3), proportion, wa.house16,
              result.name='Renters.Pct')

}
