% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/own.rent.R
\name{own.rent}
\alias{own.rent}
\title{Describe households as renters or owners}
\usage{
own.rent(house)
}
\arguments{
\item{house}{a data frame of PUMS housing data}
}
\value{
a factor column of recoded tenure
}
\description{
In the ACS PUMS data, housing tenure is described with a 4-level
integer-coded categorical that uses NA to mark vacant units.
This function recodes tenure to a 3-level factor with no NA's.
The levels are 'Own', 'Rent' and 'Vacant/GQ'.
}
\examples{
wa.house16$Tenure <- own.rent(wa.house16)
group.count(wa.house16, 'Tenure')

}
