% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{proportion}
\alias{proportion}
\title{Estimates a Proportion from the ACS PUMS Data}
\usage{
proportion(numerator, denominator, wt.rep.num = NULL)
}
\arguments{
\item{numerator}{data frame with the records of the subset class (numerator)}

\item{denominator}{data frame with records of the population (denominator)}

\item{wt.rep.num}{the weight replicate number, if a replicate weight}
}
\value{
The point estimate of the proportion
}
\description{
Estimates a proportion from the ACS PUMS data (either housing or population) 
using the sample weights.
}
\references{
\url{http://www2.census.gov/programs-surveys/acs/tech_docs/pums/ACS2013_PUMS_README.pdf}
}

