% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.frame.R
\name{quantile.frame}
\alias{quantile.frame}
\title{Compute quantiles and standard errors of quantiles on ACS PUMS data}
\usage{
\method{quantile}{frame}(x, result.name, field, probs = c(0.1, 0.25, 0.5,
  0.75, 0.9))
}
\arguments{
\item{x}{a data frame of PUMS data}

\item{result.name}{name of estimate column in result}

\item{field}{(string) name of a numerical field to take the quantiles over}

\item{probs}{a vector of quantiles to take; decimals in \code{[0.0, 1.0]}}
}
\value{
data.frame containing point estimate, standard error, and 90%
margin-of-error for the selected quantiles of \code{x$field}
}
\description{
Estimate a range of quantiles on ACS PUMS data using the sample weights for the
point estimate and the direct method, involving the replicate weights for
estimating the standard errors of those quantiles.
}
\examples{
# Some quantiles of household income for Washington State in 2016
quantile.frame(wa.house16, 'HH.Inc', 'HINCP')

}
