% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rent.burden.R
\name{rent.burden}
\alias{rent.burden}
\title{Define rent burden of households.}
\usage{
rent.burden(house)
}
\arguments{
\item{house}{a data frame of PUMS housing data}
}
\value{
factor column with levels of rent burden for households
}
\description{
Households are said to be rent-burdened if their rent is more than 30%
of household income.
This function cuts the variable \code{GRPIP} in the housing data,
which is gross rent as a percentage of household income, into bins
for rent below 30%, 30% to 40%, 40% to 50% and above 50% of income.
}
\examples{
wa.house16$Rent.Burden <- rent.burden(wa.house16)
group.count(wa.house16, 'Rent.Burden', drop.na.group=TRUE)

}
