% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.acs.mean.R
\name{se.acs.mean}
\alias{se.acs.mean}
\title{Estimates standard error of the mean for ACS PUMS data.}
\usage{
se.acs.mean(x, data.field)
}
\arguments{
\item{x}{a data.frame or similar with data filtered to relevant rows}

\item{data.field}{the column of x to take the mean of}
}
\value{
The standard error for the mean of data.field for this data
}
\description{
Computes a standard error of the mean of a field by 
the direct method from the ACS documents.
}
\references{
\url{http://www2.census.gov/programs-surveys/acs/tech_docs/pums/ACS2013_PUMS_README.pdf}
}
