% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.estimate.R
\name{se.estimate}
\alias{se.estimate}
\title{Compute Standard Errors on ACS PUMS Data}
\usage{
se.estimate(x)
}
\arguments{
\item{x}{a data frame or similar with data filtered to relevant rows.}
}
\value{
The standard error for the population total of this data.
}
\description{
Computes standard errors for count estimates from the ACS 
PUMS data by the direct method, using replicate weights.
}
\references{
\url{http://www2.census.gov/programs-surveys/acs/tech_docs/pums/ACS2013_PUMS_README.pdf}
}

