% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.proportion.R
\name{se.proportion}
\alias{se.proportion}
\title{Compute Standard Errors for Proportions from the ACS PUMS Data}
\usage{
se.proportion(numerator, denominator)
}
\arguments{
\item{numerator}{data frame with the records of the subset class (numerator)}

\item{denominator}{data frame with records of the population (denominator)}
}
\value{
The standard error for the proportion
}
\description{
Estimates the standard error for a proportion 
from the ACS PUMS data (either housing or population) 
using direct estimation method involving the replicate weights.
}
\references{
\url{http://www2.census.gov/programs-surveys/acs/tech_docs/pums/ACS2013_PUMS_README.pdf}
}
