% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sea.pop16}
\alias{sea.pop16}
\title{2016 ACS Population Data for Seattle}
\format{A data frame with 89 variables:
\describe{
\item{SERIALNO}{household serial number, a unique identifier. This is the join key to the housing data.}
\item{PUMA}{Public Use Microdata Area}
\item{RELP}{relationship of this person to the reference person. See references.}
\item{AGEP}{age}
\item{SEX}{sex}
\item{ESR}{Employment Status}
\item{RAC1P}{Race. See references. Coded  as 1=White, 2=Black, 3,4,5=Native, 6,7=Asian or Pacific Islander, 8=Other Race, 9=Mixed}
\item{HISP}{Hispanic Ancestry. See references. Coded 1 for "not Hispanic", anything else is Hispanic}
\item{PWGTP}{expansion weight, the number of people in the population represented by this row}
\item{pwgtp1..pwgtp80}{80 replicate weights for computing standard errors}
}}
\usage{
sea.pop16
}
\description{
Partial 2016 ACS PUMS person records for Seattle.
The geography is approximate; data is for Public Use Microdata
Areas (PUMA) 11601:11605, which approximate Seattle.
Data has 6053 rows and 89 columns
Included variables are SERIALNO, PUMA, RELP, AGEP, SEX, ESR, RAC1P, HISP,
the expansion weight (PWGTP), and the 80 replicate weights.
}
\references{
2016 ACS PUMS Data Dictionary \url{http://www.census.gov/programs-surveys/acs/technical-documentation/pums/documentation.html}
}
\keyword{datasets}
