% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.df.R
\name{str_strip}
\alias{str_strip}
\title{Strips blank characters off either end of string.
This function is used instead of stringr::str_trim becasue that fails
on non-ASCII characters, which are present in some versions of the
data dictionary.}
\usage{
str_strip(s)
}
\arguments{
\item{s}{a string}
}
\value{
s, with leading and trailing whitespace stripped.
}
\description{
Strips blank characters off either end of string.
This function is used instead of stringr::str_trim becasue that fails
on non-ASCII characters, which are present in some versions of the
data dictionary.
}
