% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wa.gold}
\alias{wa.gold}
\title{Ground Truth for 2015 ACS Washington Housing Data}
\format{A data frame with 4 variables: 
\describe{
\item{characteristic}{name of estimated characteristic in 2015 ACS}
\item{pums_est_15}{point estimate of the charcteristic}
\item{pums_se_15}{standard error of estimate}
\item{pums_moe_15}{90\% confidence interval for estimate}
}}
\source{
\url{http://www.census.gov/programs-surveys/acs/technical-documentation/pums/documentation.html} {PUMS Estimates for User Verification}
}
\usage{
wa.gold
}
\description{
2015 ACS PUMS ground truth point estimates and standard errors for 12 housing characteristics for Washington State.
}
\keyword{datasets}

