% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.count.R
\name{group.count}
\alias{group.count}
\title{Compute weighted size of groups in ACS PUMS data.}
\usage{
group.count(x, gp.var, result.name = NULL, drop.na.group = FALSE)
}
\arguments{
\item{x}{a data frame of PUMS data}

\item{gp.var}{name of variable in x to group data by, a string}

\item{result.name}{name of estimate column in result.
Default is 'Count'.}

\item{drop.na.group}{drop the group where \code{gp.var} is NA; default FALSE}
}
\value{
data frame of count, standard error of the count,
        and the margin of error of the count over the groups
        defined by \code{gp.var} and for the whole data.
}
\description{
Collect point estimate, standard error and 90% margin-of-error for
the size of the groups defined by \code{gp.var} and the whole data.
}
\examples{
# Count of occupied households by tenure type for Washington State in 2016
group.count(wa.house16, 'TEN', result.name='Occ.HH', drop.na.group=TRUE)

}
