% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.proportion.R
\name{group.proportion}
\alias{group.proportion}
\title{Compute proportion of group sizes on grouped ACS PUMS data.}
\usage{
group.proportion(x, gp.var, result.name = NULL, drop.na.group = TRUE)
}
\arguments{
\item{x}{a data frame of PUMS data}

\item{gp.var}{name of variable in x to group data by, a string}

\item{result.name}{name of estimate column in result
Default is "Pct.".}

\item{drop.na.group}{drop the group where \code{gp.var} is NA}
}
\value{
data frame of proportion, standard error of the proportion,
        and the margin of error of the proportion of the sizes of
        the groups defined by \code{gp.var}.
}
\description{
Collect point estimate, standard error and 90% margin-of-error for
the proportion of sizes of the groups defined by \code{gp.var}.
}
\examples{
# Fraction of occupied households by tenure type for Washington State in 2016
group.proportion(wa.house16, 'TEN', result.name='HH.Pct', drop.na.group=TRUE)

}
