% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.total.R
\name{group.total}
\alias{group.total}
\title{Compute aggregate of numerical column on grouped ACS PUMS data.}
\usage{
group.total(x, gp.var, tot.var, result.name = "Total", drop.na.group = TRUE)
}
\arguments{
\item{x}{a data frame of PUMS data. If x is already grouped, then this
function optionally adds a level to the grouping.}

\item{gp.var}{name of variable in x to group data by (string) or NULL
if x is grouped and you do not want to add a level of grouping.
May not be NULL if x is not grouped.}

\item{tot.var}{(string) name of a numerical column in x to aggregate}

\item{result.name}{name of estimate column in result.
Default is 'Total.}

\item{drop.na.group}{drop the group where \code{gp.var} is NA; default TRUE}
}
\value{
data frame of aggregate, standard error of the aggregate,
        and the margin of error of the aggregate of \code{tot.var}
        over the groups defined by \code{gp.var} and for the whole data.
}
\description{
Collect point estimate, standard error and 90% margin-of-error for
aggregate of \code{tot.var} across groups defined by \code{gp.var} and
over the whole data.
}
\examples{
# Total household population of Washington State in 2016
group.total(wa.house16, 'TEN', 'NP', result.name='HH.Pop', drop.na.group=TRUE)

}
