% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.estimate.R
\name{line.estimate}
\alias{line.estimate}
\title{Compute an estimate and standard error from ACS PUMS data.}
\usage{
line.estimate(result.name, f, x, ...)
}
\arguments{
\item{result.name}{name of estimate column in result}

\item{f}{a function to calculate the statistic.
It must take data and a weight replicate number called wt.rep.num
with a default value of NULL.}

\item{x}{a data frame of PUMS data.}

\item{...}{other data passed to f}
}
\value{
the point estimate, standard error and 90% margin of error
for the size of this sample.
}
\description{
Compute point estimate, standard error and 90% margin of error
of any statistic on ACS PUMS data using the direct method,
involving replicate weights and return results on one line.
}
\examples{
# Number of households in Washington State in 2016
line.estimate('HH.Units', estimate, wa.house16)

# Fraction of Washington State households that rent for cash
line.estimate('Renters.Pct', proportion, subset(wa.house16, TEN==3), wa.house16)

}
